requirements.txt
Pillow==8.0.1
pandas==1.1.4
urllib3==1.26.2
jupyter
ipython
opencv-python==4.1.2.30
matplotlib